requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
passlib[bcrypt]>=1.7.0
python-jose[cryptography]>=3.0.0
httpx>=0.24.0
